+++ src/main/java/com/epam/reportportal/cucumber/AbstractReporter.java
@@ -15,43 +15,66 @@
  */
 package com.epam.reportportal.cucumber;
 
+import com.epam.reportportal.listeners.ItemStatus;
 import com.epam.reportportal.listeners.ListenerParameters;
+import com.epam.reportportal.message.ReportPortalMessage;
 import com.epam.reportportal.service.Launch;
 import com.epam.reportportal.service.ReportPortal;
+import com.epam.reportportal.service.item.TestCaseIdEntry;
+import com.epam.reportportal.service.tree.TestItemTree;
 import com.epam.reportportal.utils.properties.SystemAttributesExtractor;
 import com.epam.ta.reportportal.ws.model.FinishExecutionRQ;
 import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import com.epam.ta.reportportal.ws.model.attribute.ItemAttributesRQ;
 import com.epam.ta.reportportal.ws.model.launch.StartLaunchRQ;
-import com.epam.ta.reportportal.ws.model.log.SaveLogRQ.File;
 import gherkin.formatter.Formatter;
 import gherkin.formatter.Reporter;
 import gherkin.formatter.model.*;
 import io.reactivex.Maybe;
+import org.apache.tika.Tika;
+import org.apache.tika.mime.MediaType;
+import org.apache.tika.mime.MimeTypeException;
+import org.apache.tika.mime.MimeTypes;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import rp.com.google.common.base.Supplier;
 import rp.com.google.common.base.Suppliers;
+import rp.com.google.common.io.ByteSource;
 
+import javax.annotation.Nonnull;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.IntStream;
 
 import static com.epam.reportportal.cucumber.Utils.extractAttributes;
 import static com.epam.reportportal.cucumber.Utils.getCodeRef;
+import static com.epam.reportportal.cucumber.util.ItemTreeUtils.createKey;
+import static com.epam.reportportal.cucumber.util.ItemTreeUtils.retrieveLeaf;
+import static java.util.Optional.ofNullable;
 import static rp.com.google.common.base.Strings.isNullOrEmpty;
+import static rp.com.google.common.base.Throwables.getStackTraceAsString;
 
 /**
  * Abstract Cucumber formatter/reporter for Report Portal
  *
  * @author Sergey Gvozdyukevich
  * @author Andrei Varabyeu
+ * @author Vadzim Hushchanskou
  */
 public abstract class AbstractReporter implements Formatter, Reporter {
 
+	private static final Logger LOGGER = LoggerFactory.getLogger(AbstractReporter.class);
 	private static final String AGENT_PROPERTIES_FILE = "agent.properties";
-	protected static final String COLON_INFIX = ": ";
-	private static final String SKIPPED_ISSUE_KEY = "skippedIssue";
+	private static final int DEFAULT_CAPACITY = 16;
 
-	protected final ThreadLocal<String> currentFeatureUri = new ThreadLocal<>();
+	public static final TestItemTree ITEM_TREE = new TestItemTree();
+	private static volatile ReportPortal REPORT_PORTAL = ReportPortal.builder().build();
+
+	protected static final String COLON_INFIX = ": ";
+	protected static final String SKIPPED_ISSUE_KEY = "skippedIssue";
+
 	protected final ThreadLocal<RunningContext.FeatureContext> currentFeatureContext = new ThreadLocal<>();
 	protected final ThreadLocal<RunningContext.ScenarioContext> currentScenarioContext = new ThreadLocal<>();
 
@@ -93,6 +116,14 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 		}
 	});
 
+	public static ReportPortal getReportPortal() {
+		return REPORT_PORTAL;
+	}
+
+	protected static void setReportPortal(ReportPortal reportPortal) {
+		REPORT_PORTAL = reportPortal;
+	}
+
 	/**
 	 * Extension point to customize ReportPortal instance
 	 *
@@ -111,6 +142,72 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 		launch.get().finish(finishLaunchRq);
 	}
 
+	private void addToTree(RunningContext.FeatureContext featureContext, RunningContext.ScenarioContext scenarioContext) {
+		retrieveLeaf(featureContext.getUri(), ITEM_TREE).ifPresent(suiteLeaf -> suiteLeaf.getChildItems()
+				.put(createKey(scenarioContext.getLine()), TestItemTree.createTestItemLeaf(scenarioContext.getId(), DEFAULT_CAPACITY)));
+	}
+
+	private void addToTree(RunningContext.FeatureContext context) {
+		ITEM_TREE.getTestItems().put(createKey(context.getUri()), TestItemTree.createTestItemLeaf(context.getId(), DEFAULT_CAPACITY));
+	}
+
+	/**
+	 * Start Cucumber Feature (if not started) and Scenario
+	 *
+	 * @param scenario         Scenario
+	 * @param outlineIteration - suffix to append to scenario name, can be null
+	 */
+	protected void beforeScenario(Scenario scenario, String outlineIteration) {
+		// start Feature here, because it should be started only if at least one Scenario is included.
+		// By this reason, it cannot be started in #beforeFeature method,
+		// because it will be executed even if all Scenarios in the Feature are excluded.
+		RunningContext.FeatureContext featureContext = currentFeatureContext.get();
+		if (null == featureContext.getId()) {
+			StartTestItemRQ startFeatureRq = featureContext.getItemRq();
+			Optional<Maybe<String>> root = getRootItemId();
+			startFeatureRq.setStartTime(Calendar.getInstance().getTime());
+			Maybe<String> currentFeatureId = root.map(i -> launch.get().startTestItem(i, startFeatureRq))
+					.orElseGet(() -> launch.get().startTestItem(startFeatureRq));
+			featureContext.setId(currentFeatureId);
+			addToTree(featureContext);
+		}
+		String uri = featureContext.getUri();
+		String description = Utils.getDescription(uri);
+		String codeRef = Utils.getCodeRef(uri, scenario.getLine());
+		Launch myLaunch = launch.get();
+		Maybe<String> id = Utils.startNonLeafNode(
+				myLaunch,
+				featureContext.getId(),
+				Utils.buildName(scenario.getKeyword(), COLON_INFIX, scenario.getName()),
+				description,
+				codeRef,
+				scenario.getTags(),
+				getScenarioTestItemType()
+		);
+		RunningContext.ScenarioContext scenarioContext = getCurrentScenarioContext();
+		scenarioContext.setId(id);
+		scenarioContext.setLine(scenario.getLine());
+		scenarioContext.setFeatureUri(uri);
+		if (myLaunch.getParameters().isCallbackReportingEnabled()) {
+			addToTree(featureContext, scenarioContext);
+		}
+	}
+
+	private void removeFromTree(RunningContext.FeatureContext featureContext, RunningContext.ScenarioContext scenarioContext) {
+		retrieveLeaf(featureContext.getUri(), ITEM_TREE).ifPresent(suiteLeaf -> suiteLeaf.getChildItems()
+				.remove(createKey(scenarioContext.getLine())));
+	}
+
+	/**
+	 * Finish Cucumber scenario
+	 */
+	protected void afterScenario() {
+		RunningContext.ScenarioContext context = getCurrentScenarioContext();
+		Utils.finishTestItem(launch.get(), context.getId(), context.getStatus());
+		currentScenarioContext.set(null);
+		removeFromTree(currentFeatureContext.get(), context);
+	}
+
 	/**
 	 * Define Start Cucumber feature RQ
 	 *
@@ -118,16 +215,16 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 	 */
 	protected void beforeFeature(Feature feature) {
 		//define start feature RQ in this method, because only here we can receive Feature details
-		String uri = currentFeatureUri.get();
+		RunningContext.FeatureContext featureContext = currentFeatureContext.get();
 		String featureKeyword = feature.getKeyword();
 		String featureName = feature.getName();
 		StartTestItemRQ startFeatureRq = new StartTestItemRQ();
-		startFeatureRq.setDescription(Utils.getDescription(uri));
-		startFeatureRq.setCodeRef(getCodeRef(uri, 0));
-		startFeatureRq.setName(Utils.buildNodeName(featureKeyword, AbstractReporter.COLON_INFIX, featureName, null));
+		startFeatureRq.setDescription(Utils.getDescription(featureContext.getUri()));
+		startFeatureRq.setCodeRef(getCodeRef(featureContext.getUri(), 0));
+		startFeatureRq.setName(Utils.buildName(featureKeyword, AbstractReporter.COLON_INFIX, featureName));
 		startFeatureRq.setAttributes(extractAttributes(feature.getTags()));
 		startFeatureRq.setType(getFeatureTestItemType());
-		currentFeatureContext.set(new RunningContext.FeatureContext(startFeatureRq));
+		featureContext.setItemRq(startFeatureRq);
 	}
 
 	/**
@@ -135,7 +232,6 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 	 */
 	protected void afterFeature() {
 		RunningContext.FeatureContext currentFeature = currentFeatureContext.get();
-		currentFeatureContext.remove();
 		if (null != currentFeature && null != currentFeature.getId()) {
 			Utils.finishTestItem(launch.get(), currentFeature.getId());
 		}
@@ -150,50 +246,18 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 		return context;
 	}
 
-	/**
-	 * Start Cucumber Feature (if not started) and Scenario
-	 *
-	 * @param scenario         Scenario
-	 * @param outlineIteration - suffix to append to scenario name, can be null
-	 */
-	protected void beforeScenario(Scenario scenario, String outlineIteration) {
-		// start Feature here, because it should be started only if at least one Scenario is included.
-		// By this reason, it cannot be started in #beforeFeature method,
-		// because it will be executed even if all Scenarios in the Feature are excluded.
-		RunningContext.FeatureContext currentFeature = currentFeatureContext.get();
-		if (null == currentFeature.getId()) {
-			StartTestItemRQ startFeatureRq = currentFeature.getItemRq();
-			Maybe<String> root = getRootItemId();
-			startFeatureRq.setStartTime(Calendar.getInstance().getTime());
-			Maybe<String> currentFeatureId;
-			if (null == root) {
-				currentFeatureId = launch.get().startTestItem(startFeatureRq);
-			} else {
-				currentFeatureId = launch.get().startTestItem(root, startFeatureRq);
-			}
-			currentFeature.setId(currentFeatureId);
-		}
-		String codeRef = currentFeatureUri.get() + ":" + scenario.getLine();
-		Maybe<String> id = Utils.startNonLeafNode(
-				launch.get(),
-				currentFeature.getId(),
-				Utils.buildStatementName(scenario, null, outlineIteration),
-				currentFeatureUri.get(),
-				codeRef,
-				scenario.getTags(),
-				getScenarioTestItemType()
-		);
-		RunningContext.ScenarioContext context = getCurrentScenarioContext();
-		context.setId(id);
-	}
-
-	/**
-	 * Finish Cucumber scenario
-	 */
-	protected void afterScenario() {
-		RunningContext.ScenarioContext context = getCurrentScenarioContext();
-		Utils.finishTestItem(launch.get(), context.getId(), context.getStatus());
-		currentScenarioContext.set(null);
+	protected StartTestItemRQ buildStartStepRequest(Step step, String stepPrefix, Match match) {
+		StartTestItemRQ rq = new StartTestItemRQ();
+		rq.setName(Utils.buildName(stepPrefix, step.getKeyword(), step.getName()));
+		rq.setDescription(Utils.buildMultilineArgument(step));
+		rq.setStartTime(Calendar.getInstance().getTime());
+		rq.setType("STEP");
+		String codeRef = Utils.getCodeRef(match);
+		rq.setParameters(Utils.getParameters(codeRef, match));
+		rq.setCodeRef(codeRef);
+		rq.setTestCaseId(ofNullable(Utils.getTestCaseId(match, codeRef)).map(TestCaseIdEntry::getId).orElse(null));
+		rq.setAttributes(Utils.getAttributes(match));
+		return rq;
 	}
 
 	/**
@@ -202,28 +266,74 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 	 * @param step  Step object
 	 * @param match Match object
 	 */
-	protected abstract void beforeStep(Step step, Match match);
+	protected void beforeStep(Step step, Match match) {
+		RunningContext.ScenarioContext context = getCurrentScenarioContext();
+		StartTestItemRQ rq = buildStartStepRequest(step, context.getStepPrefix(), match);
+		Launch myLaunch = launch.get();
+		Maybe<String> stepId = myLaunch.startTestItem(context.getId(), rq);
+		context.setCurrentStepId(stepId);
+		String stepText = step.getName();
+		context.setCurrentText(stepText);
+
+		if (myLaunch.getParameters().isCallbackReportingEnabled()) {
+			addToTree(context, stepText, stepId);
+		}
+	}
 
 	/**
 	 * Finish Cucumber step
 	 *
 	 * @param result Step result
 	 */
-	protected abstract void afterStep(Result result);
+	protected void afterStep(Result result) {
+		reportResult(result, null);
+		RunningContext.ScenarioContext context = getCurrentScenarioContext();
+		Launch myLaunch = launch.get();
+		myLaunch.getStepReporter().finishPreviousStep();
+		Utils.finishTestItem(myLaunch, context.getCurrentStepId(), Utils.mapStatus(result.getStatus()));
+		context.setCurrentStepId(null);
+		context.setCurrentText(null);
+	}
+
+	/**
+	 * Extension point to customize test creation event/request
+	 *
+	 * @param isBefore a cucumber hook type object
+	 * @return Request to ReportPortal
+	 */
+	protected StartTestItemRQ buildStartHookRequest(boolean isBefore) {
+		StartTestItemRQ rq = new StartTestItemRQ();
+		rq.setType(isBefore ? "BEFORE_TEST" : "AFTER_TEST");
+		rq.setName(isBefore ? "Before hooks" : "After hooks");
+		rq.setStartTime(Calendar.getInstance().getTime());
+		return rq;
+	}
 
 	/**
 	 * Called when before/after-hooks are started
 	 *
-	 * @param isBefore - if true, before-hook is started, if false - after-hook
+	 * @param isBefore - if true, before-hook is finished, if false - after-hook
 	 */
-	protected abstract void beforeHooks(Boolean isBefore);
+	protected void beforeHooks(boolean isBefore) {
+		StartTestItemRQ rq = buildStartHookRequest(isBefore);
+
+		RunningContext.ScenarioContext context = getCurrentScenarioContext();
+		context.setHookStepId(launch.get().startTestItem(context.getId(), rq));
+		context.setHookStatus(ItemStatus.PASSED);
+	}
 
 	/**
 	 * Called when before/after-hooks are finished
 	 *
 	 * @param isBefore - if true, before-hook is finished, if false - after-hook
 	 */
-	protected abstract void afterHooks(Boolean isBefore);
+	protected void afterHooks(Boolean isBefore) {
+		RunningContext.ScenarioContext context = getCurrentScenarioContext();
+		Launch myLaunch = launch.get();
+		myLaunch.getStepReporter().finishPreviousStep();
+		Utils.finishTestItem(myLaunch, context.getHookStepId(), context.getHookStatus());
+		context.setHookStepId(null);
+	}
 
 	/**
 	 * Called when a specific before/after-hook is finished
@@ -232,7 +342,10 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 	 * @param result   Hook result
 	 * @param isBefore - if true, before-hook, if false - after-hook
 	 */
-	protected abstract void hookFinished(Match match, Result result, Boolean isBefore);
+	protected void hookFinished(Match match, Result result, Boolean isBefore) {
+		reportResult(result, (isBefore ? "Before" : "After") + " hook: " + match.getLocation());
+		getCurrentScenarioContext().setHookStatus(Utils.mapStatus(result.getStatus()));
+	}
 
 	/**
 	 * Report test item result and error (if present)
@@ -243,13 +356,15 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 	protected void reportResult(Result result, String message) {
 		String cukesStatus = result.getStatus();
 		String level = Utils.mapLevel(cukesStatus);
+		if (message != null) {
+			Utils.sendLog(message, level);
+		}
 		String errorMessage = result.getErrorMessage();
 		if (errorMessage != null) {
-			Utils.sendLog(errorMessage, level, null);
+			Utils.sendLog(errorMessage, level);
 		}
-
-		if (message != null) {
-			Utils.sendLog(message, level, null);
+		if (result.getError() != null) {
+			Utils.sendLog(getStackTraceAsString(result.getError()), level);
 		}
 		RunningContext.ScenarioContext currentScenario = getCurrentScenarioContext();
 		currentScenario.updateStatus(Utils.mapStatus(result.getStatus()));
@@ -260,6 +375,7 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 	 *
 	 * @return test item name
 	 */
+	@Nonnull
 	protected abstract String getFeatureTestItemType();
 
 	/**
@@ -267,6 +383,7 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 	 *
 	 * @return test item name
 	 */
+	@Nonnull
 	protected abstract String getScenarioTestItemType();
 
 	/********************************
@@ -296,18 +413,45 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 		beforeStep(getCurrentScenarioContext().getNextStep(), match);
 	}
 
+	private static final ThreadLocal<Tika> TIKA_THREAD_LOCAL = ThreadLocal.withInitial(Tika::new);
+
+	private volatile MimeTypes mimeTypes = null;
+
+	private MimeTypes getMimeTypes() {
+		if (mimeTypes == null) {
+			mimeTypes = MimeTypes.getDefaultMimeTypes();
+		}
+		return mimeTypes;
+	}
+
+	/**
+	 * Send a log with data attached.
+	 *
+	 * @param mimeType an attachment type
+	 * @param data     data to attach
+	 */
 	@Override
 	public void embedding(String mimeType, byte[] data) {
-		File file = new File();
-		file.setName(UUID.randomUUID().toString());
-		file.setContent(data);
-		file.setContentType(mimeType);
-		Utils.sendLog("embedding", "UNKNOWN", file);
+		String type = mimeType;
+		try {
+			type = TIKA_THREAD_LOCAL.get().detect(new ByteArrayInputStream(data));
+		} catch (IOException e) {
+			// nothing to do we will use bypassed mime type
+			LOGGER.warn("Mime-type not found", e);
+		}
+		String prefix = "";
+		try {
+			MediaType mt = getMimeTypes().forName(type).getType();
+			prefix = mt.getType();
+		} catch (MimeTypeException e) {
+			LOGGER.warn("Mime-type not found", e);
+		}
+		ReportPortal.emitLog(new ReportPortalMessage(ByteSource.wrap(data), type, prefix), "UNKNOWN", Calendar.getInstance().getTime());
 	}
 
 	@Override
 	public void write(String text) {
-		Utils.sendLog(text, "INFO", null);
+		Utils.sendLog(text, "INFO");
 	}
 
 	@Override
@@ -317,8 +461,9 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 
 	@Override
 	public void uri(String uri) {
-		currentFeatureUri.set(uri);
-		launch.get().start();
+		currentFeatureContext.set(new RunningContext.FeatureContext(uri));
+		Maybe<String> launchId = launch.get().start();
+		ITEM_TREE.setLaunchId(launchId);
 	}
 
 	@Override
@@ -399,6 +544,21 @@ public abstract class AbstractReporter implements Formatter, Reporter {
 		afterFeature();
 	}
 
-	protected abstract Maybe<String> getRootItemId();
+	protected abstract Optional<Maybe<String>> getRootItemId();
 
+	protected void addToTree(RunningContext.ScenarioContext scenarioContext, String text, Maybe<String> stepId) {
+		retrieveLeaf(
+				scenarioContext.getFeatureUri(),
+				scenarioContext.getLine(),
+				ITEM_TREE
+		).ifPresent(scenarioLeaf -> scenarioLeaf.getChildItems().put(createKey(text), TestItemTree.createTestItemLeaf(stepId, 0)));
+	}
+
+	protected void removeFromTree(RunningContext.ScenarioContext scenarioContext, String text) {
+		retrieveLeaf(
+				scenarioContext.getFeatureUri(),
+				scenarioContext.getLine(),
+				ITEM_TREE
+		).ifPresent(scenarioLeaf -> scenarioLeaf.getChildItems().remove(createKey(text)));
+	}
 }