+++ src/main/java/com/epam/reportportal/cucumber/ScenarioReporter.java
@@ -15,11 +15,8 @@
  */
 package com.epam.reportportal.cucumber;
 
-import com.epam.reportportal.listeners.ItemStatus;
-import com.epam.reportportal.service.Launch;
 import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import gherkin.formatter.model.Match;
-import gherkin.formatter.model.Result;
 import gherkin.formatter.model.Step;
 import io.reactivex.Maybe;
 import org.slf4j.Logger;
@@ -27,7 +24,9 @@ import org.slf4j.LoggerFactory;
 import rp.com.google.common.base.Supplier;
 import rp.com.google.common.base.Suppliers;
 
+import javax.annotation.Nonnull;
 import java.util.Calendar;
+import java.util.Optional;
 
 /**
  * Cucumber reporter for ReportPortal that reports scenarios as test methods.
@@ -64,72 +63,40 @@ public class ScenarioReporter extends AbstractReporter {
 	});
 
 	@Override
-	protected void beforeStep(Step step, Match match) {
-		RunningContext.ScenarioContext context = getCurrentScenarioContext();
-		StartTestItemRQ rq = Utils.buildStartStepRequest(context.getStepPrefix(), step, match, false);
+	protected StartTestItemRQ buildStartStepRequest(Step step, String stepPrefix, Match match) {
+		StartTestItemRQ rq = super.buildStartStepRequest(step, stepPrefix, match);
 		rq.setHasStats(false);
-		context.setCurrentStepId(launch.get().startTestItem(context.getId(), rq));
+		return rq;
 	}
 
 	@Override
-	protected void afterStep(Result result) {
-		reportResult(result, null);
-		RunningContext.ScenarioContext context = getCurrentScenarioContext();
-		Launch myLaunch = launch.get();
-		Utils.finishTestItem(myLaunch, context.getCurrentStepId(), Utils.mapStatus(result.getStatus()));
-		context.setCurrentStepId(null);
-		myLaunch.getStepReporter().finishPreviousStep();
-	}
-
-	@Override
-	protected void beforeHooks(Boolean isBefore) {
-		StartTestItemRQ rq = new StartTestItemRQ();
+	protected StartTestItemRQ buildStartHookRequest(boolean isBefore) {
+		StartTestItemRQ rq = super.buildStartHookRequest(isBefore);
 		rq.setHasStats(false);
-
-		rq.setName(isBefore ? "Before hooks" : "After hooks");
-		rq.setStartTime(Calendar.getInstance().getTime());
-		rq.setType(isBefore ? "BEFORE_TEST" : "AFTER_TEST");
-
-		RunningContext.ScenarioContext context = getCurrentScenarioContext();
-		context.setHookStepId(launch.get().startTestItem(context.getId(), rq));
-		context.setHookStatus(ItemStatus.PASSED);
+		return rq;
 	}
 
 	@Override
-	protected void afterHooks(Boolean isBefore) {
-		RunningContext.ScenarioContext context = getCurrentScenarioContext();
-		Launch myLaunch = launch.get();
-		Utils.finishTestItem(myLaunch, context.getHookStepId(), context.getHookStatus());
-		context.setHookStepId(null);
-		myLaunch.getStepReporter().finishPreviousStep();
-	}
-
-	@Override
-	protected void hookFinished(Match match, Result result, Boolean isBefore) {
-		reportResult(result, null);
-		if (result.getStatus().equals("failed")) {
-			getCurrentScenarioContext().setHookStatus(ItemStatus.FAILED);
-		}
-	}
-
-	@Override
+	@Nonnull
 	protected String getFeatureTestItemType() {
 		return RP_TEST_TYPE;
 	}
 
 	@Override
+	@Nonnull
 	protected String getScenarioTestItemType() {
 		return RP_STEP_TYPE;
 	}
 
 	@Override
-	protected Maybe<String> getRootItemId() {
-		return rootSuiteId.get();
+	@Nonnull
+	protected Optional<Maybe<String>> getRootItemId() {
+		return Optional.of(rootSuiteId.get());
 	}
 
 	@Override
 	protected void afterLaunch() {
-		if (currentFeatureUri.get() == null) {
+		if (currentFeatureContext.get() == null || currentFeatureContext.get().getId() == null) {
 			LOGGER.debug("There is no scenarios in the launch");
 			return;
 		}