+++ src/test/java/com/epam/reportportal/cucumber/ParameterStepReporterTest.java
@@ -3,7 +3,6 @@ package com.epam.reportportal.cucumber;
 import com.epam.reportportal.cucumber.integration.TestScenarioReporter;
 import com.epam.reportportal.cucumber.integration.TestStepReporter;
 import com.epam.reportportal.cucumber.integration.util.TestUtils;
-import com.epam.reportportal.listeners.ListenerParameters;
 import com.epam.reportportal.service.ReportPortal;
 import com.epam.reportportal.service.ReportPortalClient;
 import com.epam.reportportal.util.test.CommonUtils;
@@ -16,8 +15,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -49,17 +46,23 @@ public class ParameterStepReporterTest {
 
 	}
 
+	@CucumberOptions(features = "src/test/resources/features/BasicInlineParameters.feature", glue = {
+			"com.epam.reportportal.cucumber.integration.feature" }, plugin = { "pretty",
+			"com.epam.reportportal.cucumber.integration.TestStepReporter" })
+	public static class InlineParametersTestStepReporter extends AbstractTestNGCucumberTests {
+
+	}
+
 	private final String launchId = CommonUtils.namedId("launch_");
 	private final String suiteId = CommonUtils.namedId("suite_");
 	private final List<String> testIds = Stream.generate(() -> CommonUtils.namedId("test_")).limit(3).collect(Collectors.toList());
-	private final List<Pair<String, ? extends Collection<String>>> tests = testIds.stream()
+	private final List<Pair<String, List<String>>> tests = testIds.stream()
 			.map(id -> Pair.of(id, Stream.generate(() -> CommonUtils.namedId("step_")).limit(3).collect(Collectors.toList())))
 			.collect(Collectors.toList());
 
 	private final ReportPortalClient client = mock(ReportPortalClient.class);
-	private final ListenerParameters parameters = TestUtils.standardParameters();
 	private final ExecutorService executorService = Executors.newSingleThreadExecutor();
-	private final ReportPortal reportPortal = ReportPortal.create(client, parameters, executorService);
+	private final ReportPortal reportPortal = ReportPortal.create(client, TestUtils.standardParameters(), executorService);
 
 	@BeforeEach
 	public void initLaunch() {
@@ -68,29 +71,22 @@ public class ParameterStepReporterTest {
 		TestStepReporter.RP.set(reportPortal);
 	}
 
-	public static final List<Pair<String, Object>> PARAMETERS = Arrays.asList(
-			Pair.of("String", "first"),
-			Pair.of("int", 123),
-			Pair.of("String", "second"),
-			Pair.of("int", 12345),
-			Pair.of("String", "third"),
-			Pair.of("int", 12345678)
-	);
+	@SuppressWarnings("rawtypes")
+	public static final Pair[] PARAMETERS = new Pair[] { Pair.of("java.lang.String", "first"), Pair.of("int", String.valueOf(123)),
+			Pair.of("java.lang.String", "second"), Pair.of("int", String.valueOf(12345)), Pair.of("java.lang.String", "third"),
+			Pair.of("int", String.valueOf(12345678)) };
 
-	public static final List<String> STEP_NAMES = Arrays.asList(
-			String.format("When I have parameter %s", PARAMETERS.get(0).getValue()),
-			String.format("Then I emit number %s on level info", PARAMETERS.get(1).getValue().toString()),
-			String.format("When I have parameter %s", PARAMETERS.get(2).getValue()),
-			String.format("Then I emit number %s on level info", PARAMETERS.get(3).getValue().toString()),
-			String.format("When I have parameter %s", PARAMETERS.get(4).getValue()),
-			String.format("Then I emit number %s on level info", PARAMETERS.get(5).getValue().toString())
-	);
+	public static final String[] STEP_NAMES = new String[] { String.format("When I have parameter %s", PARAMETERS[0].getValue()),
+			String.format("Then I emit number %s on level info", PARAMETERS[1].getValue()),
+			String.format("When I have parameter %s", PARAMETERS[2].getValue()),
+			String.format("Then I emit number %s on level info", PARAMETERS[3].getValue()),
+			String.format("When I have parameter %s", PARAMETERS[4].getValue()),
+			String.format("Then I emit number %s on level info", PARAMETERS[5].getValue()) };
 
 	@Test
 	public void verify_agent_retrieves_parameters_from_request() {
 		TestUtils.runTests(RunOutlineParametersTestStepReporter.class);
 
-		verify(client, times(1)).startTestItem(any());
 		verify(client, times(3)).startTestItem(same(suiteId), any());
 		ArgumentCaptor<StartTestItemRQ> captor = ArgumentCaptor.forClass(StartTestItemRQ.class);
 		verify(client, times(5)).startTestItem(same(testIds.get(0)), captor.capture());
@@ -101,18 +97,19 @@ public class ParameterStepReporterTest {
 		List<StartTestItemRQ> filteredItems = items.stream()
 				.filter(i -> "STEP".equals(i.getType()) && !i.getName().startsWith("Given"))
 				.collect(Collectors.toList());
-		IntStream.range(1, filteredItems.size()).mapToObj(i -> Pair.of(i, filteredItems.get(i))).forEach(e -> {
-			StartTestItemRQ step = e.getValue();
-			assertThat(step, notNullValue());
-			String expectedName = STEP_NAMES.get(e.getKey());
-			assertThat(step.getName(), equalTo(expectedName));
-			assertThat(step.getParameters(), allOf(notNullValue(), hasSize(1)));
-			ParameterResource param = e.getValue().getParameters().get(0);
-			Pair<String, Object> expectedParam = PARAMETERS.get(e.getKey());
+		assertThat(filteredItems, hasSize(6));
+		filteredItems.forEach(i -> assertThat(i.getParameters(), hasSize(1)));
 
-			assertThat(param.getKey(), equalTo(expectedParam.getKey()));
-			assertThat(param.getValue(), equalTo(expectedParam.getValue().toString()));
-		});
+		List<String> itemNames = filteredItems.stream().map(StartTestItemRQ::getName).collect(Collectors.toList());
+		IntStream.range(0, itemNames.size())
+				.filter(i -> i % 2 == 0)
+				.forEach(i -> assertThat(itemNames.subList(i, i + 2), containsInAnyOrder(STEP_NAMES[i], STEP_NAMES[i + 1])));
+
+		List<Pair<String, String>> params = filteredItems.stream()
+				.flatMap(i -> i.getParameters().stream())
+				.map(p -> Pair.of(p.getKey(), p.getValue()))
+				.collect(Collectors.toList());
+		assertThat(params, containsInAnyOrder(PARAMETERS));
 	}
 
 	@Test
@@ -135,5 +132,29 @@ public class ParameterStepReporterTest {
 				.collect(Collectors.toList());
 		twoParameterItems.forEach(i -> assertThat(i.getParameters(), allOf(notNullValue(), hasSize(2))));
 		oneParameterItems.forEach(i -> assertThat(i.getParameters(), allOf(notNullValue(), hasSize(1))));
+	}
+
+	@Test
+	public void verify_inline_parameters() {
+		TestUtils.runTests(InlineParametersTestStepReporter.class);
+
+		ArgumentCaptor<StartTestItemRQ> captor = ArgumentCaptor.forClass(StartTestItemRQ.class);
+		verify(client, times(5)).startTestItem(same(testIds.get(0)), captor.capture());
+
+		List<StartTestItemRQ> items = captor.getAllValues();
+		assertThat(items.get(1).getParameters(), allOf(notNullValue(), hasSize(1)));
+		ParameterResource param1 = items.get(1).getParameters().get(0);
+		assertThat(param1.getKey(), equalTo("int"));
+		assertThat(param1.getValue(), equalTo("42"));
+
+		assertThat(items.get(2).getParameters(), allOf(notNullValue(), hasSize(1)));
+		ParameterResource param2 = items.get(2).getParameters().get(0);
+		assertThat(param2.getKey(), equalTo("java.lang.String"));
+		assertThat(param2.getValue(), equalTo("string"));
+
+		assertThat(items.get(3).getParameters(), allOf(notNullValue(), hasSize(1)));
+		ParameterResource param3 = items.get(3).getParameters().get(0);
+		assertThat(param3.getKey(), equalTo("my name"));
+		assertThat(param3.getValue(), equalTo("string"));
 	}
 }