+++ src/test/java/com/epam/reportportal/cucumber/ScenarioOutlineStepReporterTest.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2020 EPAM Systems
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.epam.reportportal.cucumber;
 
 import com.epam.reportportal.cucumber.integration.TestScenarioReporter;
@@ -15,6 +31,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -22,23 +39,18 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.*;
-import static org.mockito.Mockito.any;
+import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.*;
 
-public class FeatureDescriptionTest {
+/**
+ * @author <a href="mailto:ihar_kahadouski@epam.com">Ihar Kahadouski</a>
+ */
+public class ScenarioOutlineStepReporterTest {
 
-	@CucumberOptions(features = "src/test/resources/features/belly.feature", glue = {
-			"com.epam.reportportal.cucumber.integration.feature" }, plugin = { "pretty",
-			"com.epam.reportportal.cucumber.integration.TestScenarioReporter" })
-	public static class BellyScenarioReporter extends AbstractTestNGCucumberTests {
-
-	}
-
-	@CucumberOptions(features = "src/test/resources/features/belly.feature", glue = {
+	@CucumberOptions(features = "src/test/resources/features/BasicScenarioOutlineParameters.feature", glue = {
 			"com.epam.reportportal.cucumber.integration.feature" }, plugin = { "pretty",
 			"com.epam.reportportal.cucumber.integration.TestStepReporter" })
-	public static class BellyStepReporter extends AbstractTestNGCucumberTests {
+	public static class RunOutlineParametersTestStepReporter extends AbstractTestNGCucumberTests {
 
 	}
 
@@ -49,8 +61,8 @@ public class FeatureDescriptionTest {
 			.map(id -> Pair.of(id, Stream.generate(() -> CommonUtils.namedId("step_")).limit(3).collect(Collectors.toList())))
 			.collect(Collectors.toList());
 
-	private final ListenerParameters parameters = TestUtils.standardParameters();
 	private final ReportPortalClient client = mock(ReportPortalClient.class);
+	private final ListenerParameters parameters = TestUtils.standardParameters();
 	private final ExecutorService executorService = Executors.newSingleThreadExecutor();
 	private final ReportPortal reportPortal = ReportPortal.create(client, parameters, executorService);
 
@@ -61,42 +73,17 @@ public class FeatureDescriptionTest {
 		TestStepReporter.RP.set(reportPortal);
 	}
 
-	private static final String FEATURE_CODE_REFERENCES = "src/test/resources/features/belly.feature";
-
-	private static final String SCENARIO_CODE_REFERENCES = "src/test/resources/features/belly.feature";
-
+	// Do not add iteration indexes / numbers, since it breaks re-runs
 	@Test
-	public void verify_code_reference_scenario_reporter() {
-		TestUtils.runTests(BellyScenarioReporter.class);
+	public void verify_scenario_outline_names() {
+		TestUtils.runTests(RunOutlineParametersTestStepReporter.class);
 
 		verify(client, times(1)).startTestItem(any());
 		ArgumentCaptor<StartTestItemRQ> captor = ArgumentCaptor.forClass(StartTestItemRQ.class);
-		verify(client, times(1)).startTestItem(same(suiteId), captor.capture());
-		verify(client, times(1)).startTestItem(same(testIds.get(0)), captor.capture());
+		verify(client, times(3)).startTestItem(same(suiteId), captor.capture());
 
-		List<StartTestItemRQ> items = captor.getAllValues();
+		List<String> items = captor.getAllValues().stream().map(StartTestItemRQ::getName).collect(Collectors.toList());
 
-		StartTestItemRQ feature = items.get(0);
-		StartTestItemRQ scenario = items.get(1);
-
-		assertThat(feature.getDescription(), allOf(notNullValue(), equalTo(FEATURE_CODE_REFERENCES)));
-		assertThat(scenario.getDescription(), allOf(notNullValue(), equalTo(SCENARIO_CODE_REFERENCES)));
-	}
-
-	@Test
-	public void verify_code_reference_step_reporter() {
-		TestUtils.runTests(BellyStepReporter.class);
-
-		ArgumentCaptor<StartTestItemRQ> captor = ArgumentCaptor.forClass(StartTestItemRQ.class);
-		verify(client, times(1)).startTestItem(captor.capture());
-		verify(client, times(1)).startTestItem(same(suiteId), captor.capture());
-		verify(client, times(5)).startTestItem(same(testIds.get(0)), any());
-
-		List<StartTestItemRQ> items = captor.getAllValues();
-		StartTestItemRQ feature = items.get(0);
-		StartTestItemRQ scenario = items.get(1);
-
-		assertThat(feature.getDescription(), allOf(notNullValue(), equalTo(FEATURE_CODE_REFERENCES)));
-		assertThat(scenario.getDescription(), allOf(notNullValue(), equalTo(SCENARIO_CODE_REFERENCES)));
+		assertThat(items, equalTo(Collections.nCopies(3, "Scenario Outline: Test with different parameters")));
 	}
 }