+++ src/test/java/com/epam/reportportal/cucumber/integration/feature/ParametersTest.java
@@ -1,5 +1,8 @@
 package com.epam.reportportal.cucumber.integration.feature;
 
+import com.epam.reportportal.util.test.CommonUtils;
+import cucumber.api.java.After;
+import cucumber.api.java.Before;
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;