+++ src/test/java/com/epam/reportportal/cucumber/integration/feature/ReportsTestWithParameters.java
@@ -1,5 +1,6 @@
 package com.epam.reportportal.cucumber.integration.feature;
 
+import com.epam.reportportal.annotations.ParameterKey;
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
@@ -28,5 +29,20 @@ public class ReportsTestWithParameters {
 	@Then("I emit number (\\d+) on level info")
 	public void infoLevel(int parameters) {
 		LOGGER.info("Test with parameters: " + parameters);
+	}
+
+	@Given("It is a step with an integer parameter (\\d+)")
+	public void iHaveAnIntInlineParameter(int parameter) {
+		LOGGER.info("Integer parameter: " + parameter);
+	}
+
+	@When("I have a step with a string parameter (\\w+)")
+	public void iHaveAnStrInlineParameter(String str) {
+		LOGGER.info("String parameter {}", str);
+	}
+
+	@When("I have a step with a named string parameter (\\w+)")
+	public void iHaveANamedStrInlineParameter(@ParameterKey("my name") String str) {
+		LOGGER.info("String parameter {}", str);
 	}
 }