+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Callback reporting
+### Changed
+- Test step parameters handling
+- Mime type processing for data embedding was improved
+### Fixed
+- Manually-reported nested steps now correctly fail all parents
+### Removed
+- Scenario Outline iteration number in item names, to not break re-runs
 
 ## [5.0.0-BETA-13]
 ### Fixed